Refactored cli and config
